Add windows and mac ci (#8)
* Try linux, mac and windows

* Forced int64 in labtrans

* PCHazard with int64 index

* Set workflow back to master

+++ pycox/preprocessing/label_transforms.py
@@ -133,7 +133,7 @@ class LabTransDiscreteTime:
         durations = durations.astype(self._dtype)
         events = _values_if_series(events)
         idx_durations, events = self.idu.transform(durations, events)
-        return idx_durations, events.astype('float32')
+        return idx_durations.astype('int64'), events.astype('float32')
 
     @property
     def out_features(self):
@@ -220,7 +220,7 @@ class LabTransPCHazard:
             t_frac[idx_durations == 0] = 0
             events[idx_durations == 0] = 0
         idx_durations = idx_durations - 1
-        return idx_durations, events.astype('float32'), t_frac.astype('float32')
+        return idx_durations.astype('int64'), events.astype('float32'), t_frac.astype('float32')
 
     @property
     def out_features(self):
